# fintech-auth/openapi/openapi.v1.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Digital Bank API
  license:
    name: License TBD later
  description: API for Digital Bank Authentication
  contact:
    name: Aleksey Lysak
servers:
  - url: 'http://localhost:8080'
paths:
  /auth/signup:
    post:
      summary: Initiate Sign Up
      operationId: signup-initiate
      tags:
        - signup
      responses:
        '202':
          description: Request Accepted
        '403':
          description: Request Forbidden
        '409':
          description: Request Conflict
        '500':
          description: Internal Server Error
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientData'
            examples: {}
        description: Initiate SignUp Payload
      description: Initiate Digital Bank Sign Up
    patch:
      summary: Complete Sign Up
      operationId: signup-complete
      tags:
        - signup
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtAuthToken'
        '403':
          description: Request Forbidden
        '500':
          description: Internal Server Error
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupComplete'
            examples: {}
        description: Complete SignUp Payload
      description: Complete Digital Bank Sign Up
  /auth/password/reset:
    post:
      summary: Initiate Password Reset
      operationId: password-reset-initiate
      tags:
        - password
      responses:
        '202':
          description: Request Accepted
        '403':
          description: Request Forbidden
        '500':
          description: Internal Server Error
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientData'
            examples: {}
        description: Reset Password Request Payload
      description: Reset Passsword
    parameters: []
    patch:
      summary: Complete Password Reset
      operationId: password-reset-complete
      responses:
        '200':
          description: OK
      description: Complete Reset Passsword
components:
  schemas:
    ClientData:
      description: Full client data
      type: object
      title: ClientData
      x-examples: {}
      properties:
        phone:
          type: string
          description: Client phone number
        contract:
          type: string
          description: Client contract number
        birthdate:
          type: string
          format: date
          description: Client birthday
      required:
        - phone
        - contract
        - birthdate
    JwtAuthToken:
      description: JWT Authentication Token
      type: object
      title: JwtToken
      x-examples: {}
      properties:
        accessToken:
          type: string
          description: JWT Authentication Token
        expiresIn:
          type: string
          description: Token expiration time (in seconds)
        tokenType:
          type: string
          description: Token type
    SignupComplete:
      description: ''
      type: object
      title: SignupComplete
      properties:
        phone:
          type: string
          description: Client phone number
        otpCode:
          type: string
          description: OTP Token received in SMS
        pushToken:
          type: string
          description: FCM push token
        publicKey:
          type: string
          description: Device public key
        password:
          type: string
          description: Client password
      required:
        - phone
        - otpCode
        - password
        - pushToken
        - publicKey
    ResetComplete:
      description: ''
      type: object
      title: SignupComplete
      properties:
        phone:
          type: string
          description: Client phone number
        otpCode:
          type: string
          description: OTP Token received in SMS
        pushToken:
          type: string
          description: FCM push token
        publicKey:
          type: string
          description: Device public key
        password:
          type: string
          description: Client password
      required:
        - phone
        - otpCode
        - password
        - pushToken
        - publicKey
tags:
  - name: signup
  - name: signin
  - name: device
  - name: password